requirements.txt
streamlit
streamlit_modal
google-generativeai
geopy
requests
Pillow